accuknox_social/requirements.txt
django
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-ratelimit==4.1.0
psycopg2
python-decouple==3.8